scripts/requirements.txt
snowflake-connector-python
pandas
python-dotenv
python-dateutil
cryptography
